refactor: extract routeChangeListener into own function
